+++ src/lib/app-provider/index.ts
@@ -2,3 +2,4 @@ export * from "./hooks";
 export * from "./tx";
 export * from "./contexts";
 export * from "./queries";
+export * from "./env";