+++ src/lib/services/txService.ts
@@ -5,7 +5,11 @@ import type {
 import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
 
-import { useBaseApiRoute, useCelatoneApp } from "lib/app-provider";
+import {
+  CELATONE_QUERY_KEYS,
+  useBaseApiRoute,
+  useCelatoneApp,
+} from "lib/app-provider";
 import {
   getTxsByAddressPagination,
   getTxsCountByAddress,
@@ -63,7 +67,7 @@ export const useTxData = (
   );
 
   return useQuery({
-    queryKey: ["tx_data", txsApiRoute, txHash] as string[],
+    queryKey: [CELATONE_QUERY_KEYS.TX_DATA, txsApiRoute, txHash] as string[],
     queryFn,
     enabled: enabled && Boolean(txHash && isTxHash(txHash)),
     refetchOnWindowFocus: false,
@@ -137,7 +141,7 @@ export const useTxsByAddressPagination = (
   }, [accountId, address, expression, indexerGraphClient, offset, pageSize]);
   return useQuery(
     [
-      "transactions_by_address_pagination",
+      CELATONE_QUERY_KEYS.TXS_BY_ADDRESS_PAGINATION,
       address,
       accountId,
       search,
@@ -191,7 +195,7 @@ export const useTxsCountByAddress = ({
 
   return useQuery(
     [
-      "transactions_count_by_address",
+      CELATONE_QUERY_KEYS.TXS_BY_ADDRESS_COUNT,
       address,
       search,
       filters,
@@ -243,7 +247,7 @@ export const useTxsByPoolIdPagination = (
 
   return useQuery(
     [
-      "transactions_by_pool_id",
+      CELATONE_QUERY_KEYS.POOL_TRANSACTION_BY_ID,
       poolId,
       type,
       offset,
@@ -276,7 +280,12 @@ export const useTxsCountByPoolId = (
   }, [expression, indexerGraphClient]);
 
   return useQuery(
-    ["transactions_count_by_pool_id", poolId, type, indexerGraphClient],
+    [
+      CELATONE_QUERY_KEYS.POOL_TRANSACTION_BY_ID_COUNT,
+      poolId,
+      type,
+      indexerGraphClient,
+    ],
     createQueryFnWithTimeout(queryFn, 5000),
     {
       enabled: !!poolId,
@@ -326,7 +335,7 @@ export const useTxs = (
   );
 
   return useQuery(
-    ["transaction_list", offset, pageSize, indexerGraphClient],
+    [CELATONE_QUERY_KEYS.TXS, offset, pageSize, indexerGraphClient],
     queryFn
   );
 };
@@ -341,7 +350,7 @@ export const useTxsCount = (): UseQueryResult<Option<number>> => {
     [indexerGraphClient]
   );
 
-  return useQuery(["transaction_list_count", indexerGraphClient], queryFn);
+  return useQuery([CELATONE_QUERY_KEYS.TXS_COUNT, indexerGraphClient], queryFn);
 };
 
 export const useTxsByBlockHeightPagination = (
@@ -387,7 +396,7 @@ export const useTxsByBlockHeightPagination = (
 
   return useQuery(
     [
-      "transactions_by_block_height_pagination",
+      CELATONE_QUERY_KEYS.TXS_BY_BLOCK_HEIGHT_PAGINATION,
       indexerGraphClient,
       limit,
       offset,
@@ -420,7 +429,7 @@ export const useTxsCountByBlockHeight = (
   );
 
   return useQuery(
-    ["transactions_count_by_block_height", indexerGraphClient, height],
+    [CELATONE_QUERY_KEYS.TXS_BY_BLOCK_HEIGHT_COUNT, indexerGraphClient, height],
     queryFn,
     {
       keepPreviousData: true,