+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#422](https://github.com/alleslabs/celatone-frontend/pull/422) Create query keys enum
 - [#450](https://github.com/alleslabs/celatone-frontend/pull/450) Retrieve asset prices only when necessary (/prices url path)
 - [#451](https://github.com/alleslabs/celatone-frontend/pull/451) Add jest test cases for json, date, and number in utils
 - [#412](https://github.com/alleslabs/celatone-frontend/pull/412) Refactor contract detail data loading